backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
networkx
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart